+++ cli/actions/deploy.go
@@ -52,6 +52,15 @@ func (sd *Deploy) Do() error {
 
 	d.Deployment.Version = newVersion
 
+	defer func() {
+		messages.ReportLogFieldsMessageToConsole(
+			fmt.Sprintf("\nSingularity URL : %s\n", d.Meta.Links["SingularityURL"]),
+			logging.InformationLevel,
+			sd.LogSink,
+		)
+
+	}()
+
 	updateResponse, err := updater.Update(d, sd.User.HTTPHeaders())
 	if err != nil {
 		return errors.Wrap(err, "Failed to update deployment")
@@ -100,6 +109,7 @@ func (sd *Deploy) Do() error {
 			p.Wait()
 			p.RemoveBar(bar)
 		}
+
 		return result
 	}
 	messages.ReportLogFieldsMessageToConsole(