+++ ext/singularity/deployer.go
@@ -339,6 +339,17 @@ func MakeRequestID(depID sous.DeploymentID) (string, error) {
 	return reqBase, nil
 }
 
+// MakeRequestURL creates a singularity request url
+func MakeRequestURL(baseURL string, requestID string) (string, error) {
+	if len(baseURL) == 0 {
+		return "", errors.Errorf("baseURL can not be empty : %s", baseURL)
+	}
+	if len(requestID) == 0 {
+		return "", errors.Errorf("requestID can not be empty : %s", requestID)
+	}
+	return fmt.Sprintf("%s/request/%s", baseURL, requestID), nil
+}
+
 func computeDeployID(d *sous.Deployable) string {
 	return computeDeployIDFromUUID(d, uuid.NewV4())
 }