+++ CHANGELOG.md
@@ -10,7 +10,9 @@ with respect to its command line interface and HTTP interface
 ### Changed
 * Client: when receiving bad content type print the HTTP status code and
   text along with that error message.
-
+* Both: Send the request URL back to the client and print to
+	console on deploy
+	
 ## [0.5.111](//github.com/opentable/sous/compare/0.5.110...0.5.111)
 ### Fixed
 * Client: builds using split container strategy were broken due to unique